refactor(pdf): extract shared brand primitives from the offer PDF
New components/pdf/shared.tsx holds everything brand-level for
client-side PDFs: Inter + Relative Pro Book font registration, design
tokens, Greek-uppercase helper, the vector QR code, lucide icon
strokes, and the Brand lockup (butterfly + regular-weight wordmark —
never bold, butterfly dominant). The offer PDF now consumes these.

Also extracts the repeated generate-and-download-a-Blob dance (offer
PDF button + three docx generators) to lib/utils/download.ts.
